requirements.txt
requests
pyodbc
Pillow
brother_ql
urllib3